backend/requirements.txt
# Web Framework
quart==0.19.4
hypercorn==0.17.3

# MCP HTTP Transport (Sprint 3+ HTTP Mode)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database & Cache
python-oracledb>=1.0.0
redis[hiredis]>=5.0.0

# HTTP Client
httpx[http2]>=0.25.0

# AI & Search
openai>=1.0.0
azure-search-documents>=11.4.0

# MCP Integration (Sprint 3+)
mcp>=0.9.0

# Authentication & Security
bcrypt>=4.0.0
python-multipart>=0.0.6

# Logging
structlog>=24.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
orjson>=3.9.0
python-dateutil>=2.8.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
rich>=13.0.0